# fix insertion performance issues for our `Group` structures