@yoshuawuyts 2024-04-12
Filed #175 instead, closing in favor of the reworked approach.